debug 精度

+++ src/main/java/com/money/tiger/entity/StockBasic.java
@@ -5,6 +5,8 @@ import lombok.experimental.Accessors;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Data
 @Document(collection = "stock")
 @Accessors(chain = true)
@@ -16,20 +18,20 @@ public class StockBasic {
     private String mic;
     private int type;
     private String chn;
-    private Float yield;
-    private Float pe;
-    private Float pef;
-    private Float peg;
-    private Float price;
-    private Float h52;
-    private Float l52;
-    private Float hl;
-    private Float liangbi;
-    private Float shizhi;
-    private Float huanshoulv;
-    private Float cjlrateday;
-    private Float zcrate;
-    private Float zcweek;
+    private BigDecimal yield;
+    private BigDecimal pe;
+    private BigDecimal pef;
+    private BigDecimal peg;
+    private BigDecimal price;
+    private BigDecimal h52;
+    private BigDecimal l52;
+    private BigDecimal hl;
+    private BigDecimal liangbi;
+    private BigDecimal shizhi;
+    private BigDecimal huanshoulv;
+    private BigDecimal cjlrateday;
+    private BigDecimal zcrate;
+    private BigDecimal zcweek;
     private int tag;
     private String up;
